moves mock telemetry implementation to telemetry subproject inside test
