+++ settings.gradle
@@ -87,7 +87,8 @@ List projects = [
   'test:fixtures:minio-fixture',
   'test:fixtures:old-elasticsearch',
   'test:fixtures:s3-fixture',
-  'test:logger-usage'
+  'test:logger-usage',
+  'test:telemetry'
 ]
 
 /**
