+++ test/framework/src/main/java/org/opensearch/test/OpenSearchTestCase.java
@@ -290,7 +290,6 @@ public abstract class OpenSearchTestCase extends LuceneTestCase {
 
         // Enable Netty leak detection and monitor logger for logged leak errors
         System.setProperty("io.netty.leakDetection.level", "paranoid");
-        System.setProperty("opensearch.experimental.feature.telemetry.enabled", "true");
     }
 
     protected final Logger logger = LogManager.getLogger(getClass());
